Update documentation for multisearch options
There are additional undocumented options for multisearch now (update_if and additional_attributes) that should be included in the documentation. Included them here and added small examples. related to PR #293. Note, there is still a small bug with the additional_attributes that is waiting to be fixed in PR #317 -- a workaround is noted here.