tsearch: add configuration to customize columns separator
When performing tsearch with Highlighting, the resulting query result
concatenates the searched columns with a space.

It can be hard to extract the different column data result if both
data are text (containing spaces).

In order to avoid this issue, we can allow users to customize the
columns separator (which needs to be one of the word boundaries used
by the query dictionnary).

Partial solution for #336.